refactor(clientReadyEvents): improve file cohesion
