+++ src/events/client-ready/ready.ts
@@ -1,0 +1,9 @@
+import { ClientReadyEvent } from "./clientReadyEvents.handler";
+
+const event: ClientReadyEvent = {
+  execute: (client) => {
+    console.log(`Ready! Logged in as ${client.user.tag}`);
+  },
+};
+
+export default event;
